refactor: 로컬 저장소 관리 모듈 분리
