+++ src/main.js
@@ -1,5 +1,5 @@
 import {app} from "../src/app.js";
-import {loadTodo} from "./loadTodo.js"
+import {loadTodo} from "./localStorage/loadTodo.js"
 
 app();
 loadTodo();
